requirements.txt
gdown
pycocotools
pipenv
ptflops
wget
pandas
numpy==1.23.5
opencv-python
tqdm
tensorboard
tensorboardX
pyyaml
webcolors
jsonlines
gradio
matplotlib
scipy
scikit-learn
jupyter
torch_tb_profiler
h5py
compressai @ git+https://github.com/InterDigitalInc/CompressAI.git@5b4aac0